Move function definitions into use-tile-api.ts.
